Added V_02
To fix swing rotation so it is consistent on player direction: When adding vectors together, I needed on to be relative if not a big world vector would be added to another world vector.  So made both world and relative versions of the right direction for different stages of the swing. Used world rotation for look at to fix rotation of sword at different player positions